Re-introducing "correct" large variant of logo
Our logo features an inductor that is interrupted. In Rev 1.0 of
the repeater controller, we properly used this logo. Somehow, this
got replaced by a logo that has a non-interrupted curve for the
inductor, implying a short circuit. We pulled the original logo
from the repeater controller repository and replace the incorrect
version in this commit.